requirements.txt
Package                   Version
------------------------- ------------
anyio                     3.5.0
argon2-cffi               21.3.0
argon2-cffi-bindings      21.2.0
arrow                     1.2.3
asttokens                 2.0.5
async-lru                 2.0.4
attrs                     23.1.0
Babel                     2.12.1
backcall                  0.2.0
beautifulsoup4            4.12.2
bleach                    4.1.0
certifi                   2022.12.7
cffi                      1.15.1
charset-normalizer        2.1.1
cmake                     3.25.0
comm                      0.1.4
contourpy                 1.2.0
cycler                    0.12.1
debugpy                   1.6.7
decorator                 5.1.1
defusedxml                0.7.1
entrypoints               0.4
exceptiongroup            1.0.4
executing                 0.8.3
fastjsonschema            2.16.2
filelock                  3.9.0
fonttools                 4.44.0
fqdn                      1.5.1
idna                      3.4
IProgress                 0.4
ipykernel                 6.25.0
ipython                   8.15.0
ipython-genutils          0.2.0
ipywidgets                8.1.1
isoduration               20.11.0
jedi                      0.18.1
Jinja2                    3.1.2
joblib                    1.3.2
json5                     0.9.14
jsonpointer               2.4
jsonschema                4.19.1
jsonschema-specifications 2023.7.1
jupyter                   1.0.0
jupyter_client            7.4.9
jupyter-console           6.6.3
jupyter_core              5.3.0
jupyter-events            0.7.0
jupyter-lsp               2.2.0
jupyter_server            2.7.3
jupyter_server_terminals  0.4.4
jupyterlab                4.0.6
jupyterlab-pygments       0.1.2
jupyterlab_server         2.25.0
jupyterlab-widgets        3.0.9
kiwisolver                1.4.5
lit                       15.0.7
lxml                      4.9.3
MarkupSafe                2.1.1
matplotlib                3.8.1
matplotlib-inline         0.1.6
mistune                   0.8.4
mpmath                    1.3.0
nbclassic                 0.5.5
nbclient                  0.5.13
nbconvert                 6.5.4
nbformat                  5.9.2
nest-asyncio              1.5.6
networkx                  3.0
notebook                  6.5.4
notebook_shim             0.2.2
numpy                     1.24.1
overrides                 7.4.0
packaging                 23.1
pandas                    2.1.2
pandocfilters             1.5.0
parso                     0.8.3
pexpect                   4.8.0
pickleshare               0.7.5
Pillow                    9.3.0
pip                       23.2.1
platformdirs              3.10.0
prometheus-client         0.14.1
prompt-toolkit            3.0.36
psutil                    5.9.0
ptyprocess                0.7.0
pure-eval                 0.2.2
pycparser                 2.21
Pygments                  2.15.1
pyparsing                 3.1.1
pyrsistent                0.18.0
python-dateutil           2.8.2
python-json-logger        2.0.7
pytz                      2023.3.post1
PyYAML                    6.0.1
pyzmq                     25.1.1
qtconsole                 5.4.4
QtPy                      2.4.1
referencing               0.30.2
requests                  2.31.0
rfc3339-validator         0.1.4
rfc3986-validator         0.1.1
rpds-py                   0.10.3
scikit-learn              1.3.2
scipy                     1.11.3
seaborn                   0.13.0
Send2Trash                1.8.2
setuptools                68.0.0
six                       1.16.0
sniffio                   1.2.0
soupsieve                 2.4
stack-data                0.2.0
sympy                     1.12
terminado                 0.17.1
threadpoolctl             3.2.0
tinycss2                  1.2.1
tomli                     2.0.1
torch                     2.0.1+cu118
torchaudio                2.0.2+cu118
torchsummary              1.5.1
torchvision               0.15.2+cu118
tornado                   6.3.2
tqdm                      4.66.1
traitlets                 5.7.1
triton                    2.0.0
typing_extensions         4.7.1
tzdata                    2023.3
uri-template              1.3.0
urllib3                   1.26.13
wcwidth                   0.2.5
webcolors                 1.13
webencodings              0.5.1
websocket-client          0.58.0
wheel                     0.41.2
widgetsnbextension        4.0.9
